Add OTelMetricExporter protocol and noop, console, and multiplex exporters
As part of the effort to provide an OLTP backend for Swift Metrics (see #84), we need some way to export the in-memory metrics state. This PR follows on from #89, which added the subset of the OTLP datamodel we are targeting, and provides an exporter protocol, some simple exporters (not OTLP/gRPC yet), and a multiplex convenience.

At this point, we are starting to transition more into the kinds of types spelled out in the OTel specification, which defines the names and functions for the exporter.

- Add `protocol OTelMetricExporter`.
- Add `OTelNoOpMetricExporter`.
- Add `OTelConsoleMetricExporter`.
- Add `OTelMultiplexMetricExporter`.
- Add `OTelRecordingMetricExporter` to `OTelTesting`.
- Unit tests.

There's a defined interface for exporting in-memory metrics produced by an `OTelMetricProducer`.

This paves the way for _readers_, which tie _producers_ to _exporters_, which will be coming up in a subsequent PR.