Adds support for Musl
This adds support for building the project with the "Swift Linux SDK".

For this import of Glibc needs to conditionalized for cases where `musl` is available.

Also there are some small API differences between musl and Glibc.

Also the target "OTelTesting" should be a `testTarget` otherwise it is being built together with the rest when invoking `swift build`. This fails though because the Swift Linux SDK does not conatin any testing framework.